Use PATCH and DELETE instead of new endpoints
